requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
ccxt==1.74.63
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.21.5
pandas==1.3.5
pycares==4.1.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
yarl==1.7.2
